Separate the directory processing to a separate service so that we can get transaction separation on directory-level Check if original/exported file already exists in the database and process the file only if the file has changed
